* GWorkspace/Desktop/GWDesktopWindow.m         * GWorkspace/Desktop/XBundles/XDesktopWindow/XDesktopWindow.m           more (unsigned) int to NS(U)Integer conversion
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@36074 72102866-910b-0410-8b05-ffd578937521

+++ GWorkspace/Desktop/XBundles/XDesktopWindow/XDesktopWindow.m
@@ -196,7 +196,7 @@
 }
 
 - (void)orderWindow:(NSWindowOrderingMode)place 
-         relativeTo:(int)otherWin
+         relativeTo:(NSInteger)otherWin
 {
   [super orderWindow: place relativeTo: otherWin];
   [self setLevel: NSDesktopWindowLevel];
